src/modules/cars/useCases/listAvailableCars/ListAvailableCarsUseCase.ts
import { Car } from '@modules/cars/infra/typeorm/entities/Car';
import { ICarsRepository } from '@modules/cars/repositories/ICarsRepository';
import { inject, injectable } from 'tsyringe';

interface IRequest {
  category_id?: string;
  brand?: string;
  name?: string;
}

@injectable()
class ListAvailableCarsUseCase {
  constructor(
    @inject('CarsRepository')
    private carsRepository: ICarsRepository,
  ) {}

  async execute(data: IRequest): Promise<Car[]> {
    const cars = this.carsRepository.findAvailable(data);
    return cars;
  }
}

export { ListAvailableCarsUseCase };
